Remove unique on product code and use sql null

+++ model/stock.go
@@ -4,5 +4,5 @@ type Stock struct {
 	Template
 	ProductID int64   `json:"product_id"`
 	Product   Product `gorm:"foreignkey:ProductID" json:"product"`
-	Quantity  int64   `json:"quantity"`
+	Quantity  int32   `json:"quantity"`
 }
